+++ src/constants.mjs
@@ -179,6 +179,6 @@ export const tTable = {
 // Minimum size for 95% confidence level and 5% error margin statistical significance
 export const minimumSamples = 385;
 
-export const defaultTime = 600_000_000;
+export const defaultTime = 1_000_000_000;
 
-export const defaultSamples = 128;
+export const defaultSamples = minimumSamples;